Create getMaxWaterContainer.js
This code helps find the maximum rectangular area that can be created through an array of positive integers. Each integer represents the height of the vertical-line on a chart. Find two lines which together with x-axis forms a container that would hold the greatest amount of water. Return area of water it would hold.